+++ kroxylicious-runtime/src/main/java/io/kroxylicious/proxy/internal/clusternetworkaddressconfigprovider/SniRoutingClusterNetworkAddressConfigProvider.java
@@ -6,6 +6,7 @@
 
 package io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider;
 
+import java.util.Optional;
 import java.util.Set;
 import java.util.regex.Pattern;
 
@@ -16,6 +17,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider.BrokerAddressPatternUtils.PatternAndPort;
+import io.kroxylicious.proxy.model.VirtualClusterModel;
 import io.kroxylicious.proxy.plugin.Plugin;
 import io.kroxylicious.proxy.service.ClusterNetworkAddressConfigProvider;
 import io.kroxylicious.proxy.service.ClusterNetworkAddressConfigProviderService;
@@ -23,7 +25,8 @@ import io.kroxylicious.proxy.service.HostPort;
 
 import edu.umd.cs.findbugs.annotations.NonNull;
 
-import static io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider.BrokerAddressPatternUtils.EXPECTED_TOKEN_SET;
+import static io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider.BrokerAddressPatternUtils.ALLOWED_TOKEN_SET;
+import static io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider.BrokerAddressPatternUtils.REQUIRED_TOKEN_SET;
 import static io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider.BrokerAddressPatternUtils.validatePortSpecifier;
 import static io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider.BrokerAddressPatternUtils.validateStringContainsOnlyExpectedTokens;
 import static io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider.BrokerAddressPatternUtils.validateStringContainsRequiredTokens;
@@ -37,8 +40,11 @@ import static io.kroxylicious.proxy.internal.clusternetworkaddressconfigprovider
  *    <li>{@code bootstrapAddress} a {@link HostPort} defining the host and port of the bootstrap address.</li>
  *    <li>Either {@code brokerAddressPattern} (deprecated) or ${@code advertisedBrokerAddressPattern} must be set (but not both).  These properties contain an address
  *    pattern used to advertise broker addresses.  It is addresses made from this pattern that are returned to the kafka client in the Metadata response so must be
- *    resolvable by the client.  Optionally these properties can specify a port which will be advertised to the clients.  One pattern is supported: {@code $(nodeId)}
- *    which interpolates the node id into the address.
+ *    resolvable by the client.  Optionally these properties can specify a port which will be advertised to the clients.  Two patterns ar supported:
+ *    <ul>
+ *        <li>{@code $(nodeId)} - which interpolates the node id into the address.</li>
+ *        <li>{@code $(virtualClusterName)} - which interpolates the virtual cluster name into the address.</li>
+ *    </ul>
  * </ul>
  *
  * @deprecated use {@link io.kroxylicious.proxy.config.SniHostIdentifiesNodeIdentificationStrategy} instead
@@ -49,8 +55,8 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
         ClusterNetworkAddressConfigProviderService<SniRoutingClusterNetworkAddressConfigProvider.SniRoutingClusterNetworkAddressConfigProviderConfig> {
     @NonNull
     @Override
-    public ClusterNetworkAddressConfigProvider build(SniRoutingClusterNetworkAddressConfigProviderConfig config) {
-        return new Provider(config);
+    public ClusterNetworkAddressConfigProvider build(SniRoutingClusterNetworkAddressConfigProviderConfig config, VirtualClusterModel virtualCluster) {
+        return new Provider(config, virtualCluster);
     }
 
     private static class Provider implements ClusterNetworkAddressConfigProvider {
@@ -64,12 +70,20 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
          * Creates the provider.
          *
          * @param config configuration
+         * @param virtualCluster
          */
-        private Provider(SniRoutingClusterNetworkAddressConfigProviderConfig config) {
-            this.bootstrapAddress = config.bootstrapAddress;
-            this.brokerAddressPattern = config.parsedBrokerAddressPattern;
-            this.brokerAddressNodeIdCapturingRegex = config.brokerAddressNodeIdCapturingRegex;
+        private Provider(SniRoutingClusterNetworkAddressConfigProviderConfig config, VirtualClusterModel virtualCluster) {
+            this.bootstrapAddress = new HostPort(
+                    BrokerAddressPatternUtils.replaceVirtualClusterName(config.parsedBootstrapAddressPattern, virtualCluster.getClusterName()),
+                    config.getBootstrapPort());
+            this.brokerAddressPattern = BrokerAddressPatternUtils.replaceVirtualClusterName(config.parsedBrokerAddressPattern, virtualCluster.getClusterName());
+            this.brokerAddressNodeIdCapturingRegex = buildNodeIdCaptureRegex(config, virtualCluster);
             this.advertisedPort = config.getAdvertisedPort();
+        }
+
+        private Pattern buildNodeIdCaptureRegex(SniRoutingClusterNetworkAddressConfigProviderConfig config, VirtualClusterModel virtualCluster) {
+            return Pattern.compile(BrokerAddressPatternUtils.createNodeIdCapturingRegexPattern(config.parsedBrokerAddressPattern, virtualCluster.getClusterName()),
+                    Pattern.CASE_INSENSITIVE);
         }
 
         @Override
@@ -132,14 +146,16 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
         private static final String ADVERTISED_BROKER_ADDRESS_PATTERN = "advertisedBrokerAddressPattern";
         public static final String BROKER_ADDRESS_PATTERN = "brokerAddressPattern";
 
-        private final HostPort bootstrapAddress;
+        private final String bootstrapAddress;
 
+        @JsonIgnore
+        private final String parsedBootstrapAddressPattern;
         @JsonIgnore
         private final String parsedBrokerAddressPattern;
         @JsonIgnore
-        private final Pattern brokerAddressNodeIdCapturingRegex;
-        @JsonIgnore
         private final Integer advertisedPort;
+        @JsonIgnore
+        private final Integer bootstrapPort;
 
         // present for serialize/deserialize fidelity
         @SuppressWarnings("unused")
@@ -148,7 +164,7 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
         @SuppressWarnings("unused")
         private final String advertisedBrokerAddressPattern;
 
-        public SniRoutingClusterNetworkAddressConfigProviderConfig(@JsonProperty(required = true) HostPort bootstrapAddress,
+        public SniRoutingClusterNetworkAddressConfigProviderConfig(@JsonProperty(required = true) String bootstrapAddress,
                                                                    @Deprecated(forRemoval = true, since = "0.10.0") @JsonProperty String brokerAddressPattern,
                                                                    @JsonProperty String advertisedBrokerAddressPattern) {
             this.brokerAddressPattern = brokerAddressPattern;
@@ -169,26 +185,32 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
                 brokerAddressPatternToParse = brokerAddressPattern;
             }
 
+            PatternAndPort bootstrapPatternAndPort = BrokerAddressPatternUtils.parse(bootstrapAddress);
+            Optional<Integer> bootstrapPort = bootstrapPatternAndPort.port();
+            if (bootstrapPort.isEmpty()) {
+                throw new IllegalArgumentException("bootstrapAddress " + bootstrapAddress + " does not contain a port");
+            }
+            this.bootstrapPort = bootstrapPort.get();
             PatternAndPort patternAndPort = BrokerAddressPatternUtils.parse(brokerAddressPatternToParse);
             String brokerAddressPatternPart = patternAndPort.addressPattern();
-            advertisedPort = patternAndPort.port().orElse(bootstrapAddress.port());
+            advertisedPort = patternAndPort.port().orElse(this.bootstrapPort);
 
             validatePortSpecifier(brokerAddressPatternPart, s -> {
                 throw new IllegalArgumentException(
                         brokerAddressPatternProp + " address pattern cannot have port specifier.  Found port : " + s + " within " + brokerAddressPatternPart);
             });
 
-            validateStringContainsOnlyExpectedTokens(brokerAddressPatternPart, EXPECTED_TOKEN_SET, tok -> {
+            validateStringContainsOnlyExpectedTokens(brokerAddressPatternPart, ALLOWED_TOKEN_SET, tok -> {
                 throw new IllegalArgumentException(brokerAddressPatternProp + " contains an unexpected replacement token '" + tok + "'");
             });
 
-            validateStringContainsRequiredTokens(brokerAddressPatternPart, EXPECTED_TOKEN_SET, tok -> {
+            validateStringContainsRequiredTokens(brokerAddressPatternPart, REQUIRED_TOKEN_SET, tok -> {
                 throw new IllegalArgumentException(brokerAddressPatternProp + " must contain at least one nodeId replacement pattern '" + tok + "'");
             });
 
             this.bootstrapAddress = bootstrapAddress;
+            this.parsedBootstrapAddressPattern = bootstrapPatternAndPort.addressPattern();
             this.parsedBrokerAddressPattern = brokerAddressPatternPart;
-            this.brokerAddressNodeIdCapturingRegex = BrokerAddressPatternUtils.createNodeIdCapturingRegex(brokerAddressPatternPart);
         }
 
         private static void validateOneSpecified(String brokerAddressPattern, String advertisedBrokerAddressPattern) {
@@ -200,8 +222,8 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
             }
         }
 
-        public HostPort getBootstrapAddress() {
-            return bootstrapAddress;
+        public String getBootstrapAddressPattern() {
+            return parsedBootstrapAddressPattern;
         }
 
         public String getBrokerAddressPattern() {
@@ -210,6 +232,10 @@ public class SniRoutingClusterNetworkAddressConfigProvider implements
 
         public int getAdvertisedPort() {
             return advertisedPort;
+        }
+
+        public int getBootstrapPort() {
+            return bootstrapPort;
         }
 
     }