+++ CHANGELOG.md
@@ -7,12 +7,17 @@ Format `<github issue/pr number>: <short description>`.
 
 ## SNAPSHOT
 
+* [#2185](https://github.com/kroxylicious/kroxylicious/pull/2185) Add $(virtualClusterName) placeholders to SNI bootstrap address and advertised broker address pattern
 * [#1871](https://github.com/kroxylicious/kroxylicious/issues/1871) Remove deprecated configuration property bootstrap_servers
 
 ### Changes, deprecations and removals
 
 * Remove the deprecated configuration property `bootstrap_servers` from the `targetCluster` object. Use `bootstrapServers`
   instead.
+* `virtualClusters[].gateways[].sniHostIdentifiesNode.bootstrapAddress` can now contain an optional replacement token `$(virtualClusterName)`.
+When this is present, it will be replaced with the name of that gateway's VirtualCluster.
+* `virtualClusters[].gateways[].sniHostIdentifiesNode.advertisedBrokerAddressPattern` can now contain an optional replacement toke  `$(virtualClusterName)`.
+When this is present, it will be replace with the name of that gateway's VirtualCluster.
 
 
 ## 0.12.0